moved Tile and Matrix in namespace dlaf::matrix (they are still available in namespace dlaf for compatibility reason)
