+++ include/dlaf/lapack_tile.h
@@ -19,11 +19,12 @@
 #undef I
 #endif
 
-#include "dlaf/tile.h"
+#include "dlaf/matrix/tile.h"
 #include "dlaf/types.h"
 
 namespace dlaf {
 namespace tile {
+using matrix::Tile;
 
 // See LAPACK documentation for more details.
