+++ include/dlaf/matrix/matrix_view.h
@@ -19,7 +19,7 @@
 #include "dlaf/matrix/distribution.h"
 #include "dlaf/matrix/layout_info.h"
 #include "dlaf/matrix/matrix_base.h"
-#include "dlaf/tile.h"
+#include "dlaf/matrix/tile.h"
 #include "dlaf/types.h"
 
 namespace dlaf {
@@ -96,7 +96,7 @@ MatrixView<std::add_const_t<T>, device> getConstView(MatrixType<T, device>& matr
   return getConstView(blas::Uplo::General, matrix);
 }
 
-#include "dlaf/matrix/matrix_view_const.tpp"
+}
+}
 
-}
-}
+#include "dlaf/matrix/matrix_view_const.tpp"
