+++ include/dlaf/matrix/tile.h
@@ -206,8 +206,6 @@ auto create_data(const Tile<T, device>& tile) {
                                    to_sizet(tile.size().rows()), to_sizet(tile.ld()));
 }
 
-#include <dlaf/tile.tpp>
-
 /// ---- ETI
 
 #define DLAF_TILE_ETI(KWORD, DATATYPE, DEVICE) \
@@ -225,3 +223,6 @@ DLAF_TILE_ETI(extern, std::complex<double>, Device::CPU)
 // DLAF_TILE_ETI(extern, std::complex<double>, Device::GPU)
 
 }
+}
+
+#include <dlaf/matrix/tile.tpp>
